Stop duplicating the source tree for the -nox variant.
Turns out that Emacs' habit of building things in the source tree only
applies to in-place byte-compilation of elisp files; the actual C
compilation seems to be working reasonably well in separate build
directories.

This means that all along, the real problem that I was fighting
against was debhelper's insistence on dumping both parallel builds
into debian/tmp, one clobbering the other, and that the --destdir
option to dh_auto_install was really the solution all along. It wasn't
necessary at all to cp the source tree into a different place.

So this commit stops doing that superfluous cp, and drops all usage of
the --sourcedir option that was necessary for pointing the build
systems at the different source dirs (which are now reunited).