Autogenerate debian/control from new debian/control.in
This means that the package names are only ever defined in the
changelog, and they are then parsed from there and used to construct
the control file, all the emacsVER files, and other options within the
rules file. This should make it quite easy to use this same packaging
data both for official package builds and daily snapshot builds.